Removing the PATCH response and now a 'patch' message is always returned
Motivation:
When a client sends a patch message it would get a simple string
response 'PATCHED' to imply that the server has accepted the patch. This
seemed resonable as a client would not be interested in its own patches.
But this was incorrect and the client is interested in getting the
version acknowledgement from the server, and will need to patch it's
sync engine to that any edits can be cleared.

Modifications:
- Updated the server side to no longer return the 'PATCHED' response
- Updated the js-client and test to always handle a 'patch' msg response