Adding check of client content to stringify or not.
Motivation:
Different clients will have different types of data that they insert as
documenents into the sync engines (client and server). They might opt
for pure strings or complete JavaScript objects.

Modifications:
Added a check for the type of content before performing a diff and only
stringify if the content is not of type 'string'

Result:
There are no longer any additional '"' which make the diffs inconsitent
with the server side shadows version and diffs are produced properly.