Adding support for adding a doc without any content.
Motivation:
The current version of the server only supports clients that send a
"base" document content. The demo app is an example of this which has
seed data that is sent to the server by every client.
But an equally common use case is where a client has an document id and
does not need to start with a base version as it know that someone else
created it (the user/system that provided the id).

Modifications:
The change adds support for an none existing 'content' property in the
add document message.

Result:
Both seeded and un-seed clients can be used successfully.