Clarifying unit tests
Motivation:
Some tests are a little hard to follow with regards to the expectations
in the diff-sync specification. Adding some comments to the code will
hopefully help us in the future to understand what the test do. This is
especially true for the tests that deal with restoring the client backup
shadow.

Modifications:
Updated patchRevertToBackup adding comments and a few minor refactorings
for clarity.

Result:
Will hopefully be easier for us to understand the test and make updates
in the future.

JIRA: https://issues.jboss.org/browse/AEROGEAR-1585