Moving JsonMapperTest from server-netty to core module.
Motivation:
The JsonMapperTest is currently in the server-netty module which was
because it used the ClientSyncEngine to generate PatchMessage instances
to be passed into the JsonMapper methods.

This test belongs in the core module where the JsonMapper class lives.

Modifications:
Remove the dependencies to ClientSyncEngine and instead use the builder
for DefautlEdits which allows the manual construction of diffs.

Result:
The test are located in the correct module and have no external
dependencies.