Initial refactoring to be able to support different patch/diffs
Motivation:
When trying to add a different storage option and server syncrhonizer
implementation to the server it became obvious that it needed
refactoring. The PatchMessage and the type of diff/patches
implementations that the server handles must also be provided when
adding a new storeage/syncrhonizer option.

Modifications:
PatchMessage not takes a type of subclass of Edit. The patch message is
what the client and server use to send edits. It contains a clientId and
a documentId, and a stack of Edits. The edits are the diffs, or patches,
to be applied. Some implementations might have diffs that are operations
to be performed, while others might provide a single patch to be
applied.

Result:
Hopefully is will now be possible to implement a JSON Patch or JSON
Merge implementation (and other types).