Changed Edit to have a generic type of Diff.
Motivation:
Currently an Edit has a List of Diff instances. This made sense when we
only supported DiffMatchPatch as it contains just that, a list of diffs
to be performed. But for JSON Patch and JSON Merge Patch this did not
make sense and complicated the message wire format.

Modifications:
Edit now returns a Diff which can be an type that the synchronizers need
to support.

Result:
The message wire format is simpler for JSON Patch and JSON Merge Patch.