Moving code into correct/new packages and adding package.java to them.
Motivation:
Some of the classes were in the 'org.jboss.aerogear.sync' package but
were not part of the core module. The made it harder to read the
javadocs produces as there were a lot of classed in the same packate.

Modifications:
Moved packages into packages matching there intension. Also added
package.java file to describe the new packages where needed.

Results:
Javadoc are easier to read and there is a clear separation of code.