Fixing incorrect usage of backup version in ClientSyncEngine.
Motivation:
Currently when restore of a backup shadow is done the ClientSyncEngine
would use the backups version, which on the client side represents the
servers version. It would use this version, that is the server version,
to create a new BackupShadowDocument. This is a bug.

Modifications:
The code can be simplified and instead of creating a new
BackupShadowDocument instance, the ShadowDocument that the backup holds
can be used instead.

I've also added some additional comments to the tests to clarify why
some changes were made to them.

Result:
Simpler code code and the bug does not exist anymore.

Jira: https://issues.jboss.org/browse/AGSYNC-26