ServerInMemoryDataStore should compare the server version and not client.
Motivation:
Currently, when a Edit is to be removed after a patch has been
successfully completed, the server should compare the clients stack of
edits and see if the client has acknowledged any with the current patch.
This comparision should use the Edit's serverVersion but is currently using
the clientVersion.

Modifications:
Updated the ServerInMemoryDatabase's removeEdits method to compare the
server version.