Refactoring SyncClient to be more usable.
Motivation:
Summers brought up that using Obervable and Observer is not what is
expected to be used on Android.
This commit tries to address this issue by adding our own interfaces
that and instead of having the ClientSyncEngine extend Observable it
now expectes a PatchObservable to be injected into it when it is
constructed.

JIRA:
https://issues.jboss.org/browse/AGSYNC-48